+++ test/wizard_router_test.dart
@@ -357,4 +357,36 @@ void main() {
     expect(secondPage, findsNothing);
     expect(thirdPage, findsNothing);
   });
+
+  testWidgets('has next or previous', (tester) async {
+    await pumpWizardApp(
+      tester,
+      routes: {
+        Routes.first: (_) => const Text(Routes.first),
+        Routes.second: (_) => const Text(Routes.second),
+        Routes.third: (_) => const Text(Routes.third),
+      },
+    );
+
+    final firstPage = find.text(Routes.first);
+    final wizard = Wizard.of(tester.element(firstPage));
+
+    // 1st
+    expect(wizard.hasPrevious, isFalse);
+    expect(wizard.hasNext, isTrue);
+
+    // 2nd
+    wizard.next();
+    await tester.pumpAndSettle();
+
+    expect(wizard.hasPrevious, isTrue);
+    expect(wizard.hasNext, isTrue);
+
+    // 3rd
+    wizard.next();
+    await tester.pumpAndSettle();
+
+    expect(wizard.hasPrevious, isTrue);
+    expect(wizard.hasNext, isFalse);
+  });
 }